Relocate document creation and link template creation into execution
